refractor the implementation
