+++ oci/image.go
@@ -16,21 +16,29 @@ package oci
 
 import (
 	"github.com/google/go-containerregistry/pkg/name"
-	"github.com/google/go-containerregistry/pkg/v1/remote"
+	v1 "github.com/google/go-containerregistry/pkg/v1"
+	"github.com/in-toto/go-witness/log"
 )
 
-// ResolveDigest returns the digest of the image at the reference.
-//
-// If the reference is by digest already, it simply extracts the digest.
-// Otherwise, it looks up the digest from the registry.
-func ResolveDigest(ref name.Reference, opts ...Option) (name.Digest, error) {
-	o := makeOptions(ref.Context(), opts...)
-	if d, ok := ref.(name.Digest); ok {
-		return d, nil
-	}
-	desc, err := remote.Get(ref, o.ROpt...)
-	if err != nil {
-		return name.Digest{}, err
-	}
-	return ref.Context().Digest(desc.Digest.String()), nil
+type image struct {
+	v1.Image
+	ref name.Reference
+	opt *options
+	// opt *RemoteOptions
+}
+
+func (si *image) Signatures() (Signatures, error) {
+	// Simplified signature retrieval logic
+	log.Info("image Signatures")
+	return signatures(si, si.opt)
+}
+
+func (si *image) Attestations() (Signatures, error) {
+	// Simplified attestation retrieval logic
+	log.Info("image Attestations")
+	return attestations(si, si.opt)
+}
+func (si *image) Attachment(name string) (File, error) {
+	// Simplified attachment retrieval logic
+	return nil, nil
 }
